fix(classify/symbol): don't leak memory when constructing id
Use the `get_key` method to return the symbol table key, instead of
reconstructing it and then leaking to get a `&'static str`.
To ensure safety when getting the key, the `font_econding` attribute is
made private (it's not exposed in the UI anywhere), causing the `Symbol`
to only be instantiated through the symbol table, ensuring that the key
is always valid.